refactor: 🎨 Deletion of util component used in only one file
