+++ src/bot/commands/correction.ts
@@ -11,7 +11,7 @@ import {
   TextChannel
 } from 'discord.js';
 import prisma from '../../prisma';
-import { Category, JokeTypesDescriptions, CategoriesRefs, UnsignedJokeKey } from '../../typings';
+import { Category, JokeTypesDescriptions, CategoriesRefs, UnsignedJokeKey, UnsignedJoke } from '../../typings';
 import {
   Colors,
   commandsChannelId,
@@ -33,9 +33,25 @@ import {
   showNegativeDiffs,
   showPositiveDiffs,
   info,
-  JokeCorrectionPayload,
-  findJoke
+  messageProblem,
+  tDelete,
 } from '../utils';
+import { jokeById, jokeByQuestion } from '../../controllers';
+
+enum IdType {
+  MESSAGE_ID,
+  JOKE_ID,
+  MESSAGE_QUESTION
+}
+
+export interface JokeCorrectionPayload extends UnsignedJoke {
+  id?: number;
+  correction_type: ProposalType;
+  suggestion: UnsignedJoke & {
+    message_id: string | null;
+    proposal_id: number;
+  };
+}
 
 export default class CorrectionCommand extends Command {
   constructor() {
@@ -75,7 +91,7 @@ export default class CorrectionCommand extends Command {
     interaction: ChatInputCommandInteraction<'cached'>,
     query: string
   ): Promise<JokeCorrectionPayload | null> {
-    const joke = await findJoke(interaction, query);
+    const joke = await this.findJoke(interaction, query);
     if (joke) return joke;
 
     const question = await interaction.reply({
@@ -97,7 +113,7 @@ export default class CorrectionCommand extends Command {
       });
       collector.on('collect', async (msg: Message) => {
         if (msg.deletable) setInterval(() => msg.delete().catch(() => null), 5000);
-        const joke = await findJoke(interaction, msg.content);
+        const joke = await this.findJoke(interaction, msg.content);
 
         if (joke) {
           collector.stop();
@@ -417,4 +433,139 @@ export default class CorrectionCommand extends Command {
       await message.react(reaction).catch(() => null);
     }
   }
+  
+  async findJoke(
+    interaction: ChatInputCommandInteraction<'cached'>,
+    query: string
+  ): Promise<JokeCorrectionPayload | null> {
+    const idType = this.getIdType(query);
+    if (idType === IdType.MESSAGE_ID) {
+      const proposal = await prisma.proposal.findUnique({
+        where: {
+          message_id: query
+        },
+        include: {
+          corrections: {
+            take: 1,
+            orderBy: {
+              created_at: 'desc'
+            },
+            where: {
+              merged: false,
+              refused: false
+            }
+          },
+          suggestion: {
+            include: {
+              corrections: {
+                take: 1,
+                orderBy: {
+                  created_at: 'desc'
+                },
+                where: {
+                  merged: false,
+                  refused: false
+                }
+              }
+            }
+          }
+        }
+      });
+      if (!proposal) {
+        interaction.channel
+          ?.send(
+            messageProblem(
+              `Impossible de trouver une blague ou correction liée à cet ID de blague, assurez vous que cet ID provient bien d'un message envoyé par le bot ${interaction.client.user}`
+            )
+          )
+          .then(tDelete(5000));
+        return null;
+      }
+  
+      const origin = proposal.type === ProposalType.SUGGESTION ? proposal : proposal.suggestion!;
+  
+      return {
+        id: proposal.joke_id ?? undefined,
+        type: (origin.corrections[0]?.joke_type ?? origin.joke_type) as Category,
+        joke: (origin.corrections[0]?.joke_question ?? origin.joke_question)!,
+        answer: (origin.corrections[0]?.joke_answer ?? origin.joke_answer)!,
+        correction_type: origin.merged ? ProposalType.CORRECTION : ProposalType.SUGGESTION_CORRECTION,
+        suggestion: {
+          message_id: origin.message_id,
+          proposal_id: origin.id,
+          type: origin.joke_type as Category,
+          joke: origin.joke_question!,
+          answer: origin.joke_answer!
+        }
+      };
+    }
+  
+    const joke = idType === IdType.JOKE_ID ? jokeById(Number(query)) : jokeByQuestion(query);
+    if (!joke) {
+      interaction.channel
+        ?.send(
+          messageProblem(
+            `Impossible de trouver une blague à partir de ${
+              idType === IdType.JOKE_ID ? 'cet identifiant' : 'cette question'
+            }, veuillez réessayer !`
+          )
+        )
+        .then(tDelete(5000));
+      return null;
+    }
+  
+    const proposal = await prisma.proposal.upsert({
+      create: {
+        joke_id: joke.id,
+        joke_type: joke.type,
+        joke_question: joke.joke,
+        joke_answer: joke.answer,
+        type: ProposalType.SUGGESTION,
+        merged: true
+      },
+      include: {
+        corrections: {
+          take: 1,
+          orderBy: {
+            created_at: 'desc'
+          },
+          where: {
+            merged: false,
+            refused: false
+          }
+        }
+      },
+      update: {},
+      where: {
+        joke_id: joke.id
+      }
+    });
+  
+    const correction = proposal.corrections[0];
+    return {
+      id: proposal.joke_id!,
+      type: (correction?.joke_type ?? proposal.joke_type) as Category,
+      joke: (correction?.joke_question ?? proposal.joke_question)!,
+      answer: (correction?.joke_answer ?? proposal.joke_answer)!,
+      correction_type: ProposalType.CORRECTION,
+      suggestion: {
+        message_id: proposal.message_id,
+        proposal_id: proposal.id,
+        type: proposal.joke_type as Category,
+        joke: proposal.joke_question!,
+        answer: proposal.joke_answer!
+      }
+    };
+  }
+  
+  getIdType(query: string): IdType {
+    if (isNaN(Number(query))) {
+      return IdType.MESSAGE_QUESTION;
+    }
+    if (query.length > 6) {
+      return IdType.MESSAGE_ID;
+    }
+    return IdType.JOKE_ID;
+  }
+  
 }
